refactor(cloudformation): move fargate.go into utils.go
